+++ moshi-ir/moshi-compiler-plugin/src/main/kotlin/dev/zacsweers/moshix/ir/compiler/sealed/SealedAdapterGenerator.kt
@@ -5,12 +5,15 @@ import dev.zacsweers.moshix.ir.compiler.api.AdapterGenerator
 import dev.zacsweers.moshix.ir.compiler.api.PreparedAdapter
 import dev.zacsweers.moshix.ir.compiler.labelKey
 import dev.zacsweers.moshix.ir.compiler.util.addOverride
+import dev.zacsweers.moshix.ir.compiler.util.checkIsVisible
 import dev.zacsweers.moshix.ir.compiler.util.createIrBuilder
 import dev.zacsweers.moshix.ir.compiler.util.error
 import dev.zacsweers.moshix.ir.compiler.util.generateToStringFun
 import dev.zacsweers.moshix.ir.compiler.util.irConstructorBody
 import dev.zacsweers.moshix.ir.compiler.util.irInstanceInitializerCall
+import dev.zacsweers.moshix.ir.compiler.util.irInvoke
 import dev.zacsweers.moshix.ir.compiler.util.irType
+import dev.zacsweers.moshix.ir.compiler.util.rawType
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
@@ -40,6 +43,9 @@ import org.jetbrains.kotlin.ir.declarations.IrDeclarationOrigin
 import org.jetbrains.kotlin.ir.declarations.IrFactory
 import org.jetbrains.kotlin.ir.declarations.IrField
 import org.jetbrains.kotlin.ir.declarations.IrFunction
+import org.jetbrains.kotlin.ir.declarations.IrValueParameter
+import org.jetbrains.kotlin.ir.expressions.IrCall
+import org.jetbrains.kotlin.ir.expressions.IrClassReference
 import org.jetbrains.kotlin.ir.expressions.IrConst
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
 import org.jetbrains.kotlin.ir.expressions.IrDelegatingConstructorCall
@@ -47,12 +53,16 @@ import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.expressions.IrVararg
 import org.jetbrains.kotlin.ir.expressions.impl.IrDelegatingConstructorCallImpl
 import org.jetbrains.kotlin.ir.interpreter.hasAnnotation
-import org.jetbrains.kotlin.ir.types.classFqName
+import org.jetbrains.kotlin.ir.symbols.IrClassSymbol
+import org.jetbrains.kotlin.ir.symbols.IrFunctionSymbol
+import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.classifierOrFail
 import org.jetbrains.kotlin.ir.types.createType
 import org.jetbrains.kotlin.ir.types.defaultType
+import org.jetbrains.kotlin.ir.types.getClass
 import org.jetbrains.kotlin.ir.types.impl.IrSimpleTypeImpl
 import org.jetbrains.kotlin.ir.types.typeWith
+import org.jetbrains.kotlin.ir.util.classId
 import org.jetbrains.kotlin.ir.util.constructors
 import org.jetbrains.kotlin.ir.util.defaultType
 import org.jetbrains.kotlin.ir.util.fqNameWhenAvailable
@@ -60,6 +70,8 @@ import org.jetbrains.kotlin.ir.util.getAnnotation
 import org.jetbrains.kotlin.ir.util.getSimpleFunction
 import org.jetbrains.kotlin.ir.util.isObject
 import org.jetbrains.kotlin.ir.util.packageFqName
+import org.jetbrains.kotlin.ir.util.primaryConstructor
+import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
 
@@ -78,7 +90,7 @@ private constructor(
       if (target.hasAnnotation(FqName("dev.zacsweers.moshix.sealed.annotations.NestedSealed"))) {
         target.superTypes.firstNotNullOfOrNull { supertype ->
           pluginContext
-            .referenceClass(supertype.classFqName!!)!!
+            .referenceClass(supertype.getClass()!!.classId!!)!!
             .owner
             .getAnnotation(FqName("com.squareup.moshi.JsonClass"))
             ?.labelKey()
@@ -97,12 +109,72 @@ private constructor(
           "@NestedSealed-annotated subtype ${target.fqNameWhenAvailable} is inappropriately annotated with @JsonClass(generator = " +
             "\"sealed:$labelKey\")."
         }
+        return null
       }
     }
 
-    // TODO there's no non-descriptor API for this yet
+    var fallbackStrategy: FallbackStrategy? = null
+
     val useDefaultNull =
       target.hasAnnotation(FqName("dev.zacsweers.moshix.sealed.annotations.DefaultNull"))
+
+    val fallbackAdapterAnnotation =
+      target.getAnnotation(FqName("dev.zacsweers.moshix.sealed.annotations.FallbackJsonAdapter"))
+
+    if (useDefaultNull && (fallbackAdapterAnnotation != null)) {
+      logger.error(target) {
+        "Only one of @DefaultNull or @FallbackJsonAdapter can be used at a time"
+      }
+      return null
+    }
+
+    if (fallbackAdapterAnnotation != null) {
+      val adapterType = fallbackAdapterAnnotation.getValueArgument(0) as IrClassReference
+      // TODO can we check adapter type is valid? Compiler will check it for us
+      val adapterDeclaration = adapterType.classType.rawType()
+      val constructor =
+        adapterDeclaration.primaryConstructor
+          ?: run {
+            logger.error(adapterDeclaration) {
+              "No primary constructor found for fallback adapter ${adapterDeclaration.fqNameWhenAvailable}"
+            }
+            return null
+          }
+      constructor.visibility.checkIsVisible()
+      val hasMoshiParam =
+        when (constructor.valueParameters.size) {
+          0 -> {
+            // Nothing to do
+            false
+          }
+          1 -> {
+            // Check it's a Moshi parameter
+            val moshiParam = constructor.valueParameters[0]
+            // TODO can this be simpler?
+            if (moshiSymbols.moshi != moshiParam.type) {
+              logger.error(target) {
+                "Fallback adapter type's primary constructor can only have a Moshi parameter"
+              }
+              return null
+            }
+            true
+          }
+          else -> {
+            logger.error(target) {
+              "Fallback adapter type's primary constructor can only have a Moshi parameter"
+            }
+            return null
+          }
+        }
+      fallbackStrategy =
+        FallbackStrategy.FallbackAdapter(
+          targetConstructor = constructor.symbol,
+          hasMoshiParam = hasMoshiParam
+        )
+    } else if (useDefaultNull) {
+      fallbackStrategy = FallbackStrategy.Null
+    }
+
     val objectSubtypes = mutableListOf<IrClass>()
     val seenLabels = mutableMapOf<String, IrClass>()
     var hasErrors = false
@@ -142,15 +214,12 @@ private constructor(
       return createType(
         targetType = target,
         labelKey = labelKey,
-        useDefaultNull = useDefaultNull,
+        fallbackStrategy = fallbackStrategy,
         generatedAnnotation = null,
         subtypes = sealedSubtypes,
         objectSubtypes = objectSubtypes
       )
     }
-
-    // TODO sealedParent gen?
-    //  Requires a runtime adapter! Or JsonClass(generator = "sealed-nested")
   }
 
   private fun walkTypeLabels(
@@ -283,7 +352,7 @@ private constructor(
   private fun IrFactory.createType(
     targetType: IrClass,
     labelKey: String,
-    useDefaultNull: Boolean,
+    fallbackStrategy: FallbackStrategy?,
     generatedAnnotation: IrConstructorCall?,
     subtypes: Set<Subtype>,
     objectSubtypes: List<IrClass>,
@@ -321,7 +390,7 @@ private constructor(
     adapterCls.thisReceiver = adapterReceiver
     val jsonAdapterType =
       pluginContext
-        .irType("com.squareup.moshi.JsonAdapter")
+        .irType(ClassId.fromString("com/squareup/moshi/JsonAdapter"))
         .classifierOrFail
         .typeWith(targetType.defaultType)
     adapterCls.superTypes = listOf(jsonAdapterType)
@@ -350,6 +419,31 @@ private constructor(
                   putValueArgument(1, irString(labelKey))
                 }
 
+              val moshiParam = ctor.valueParameters[0]
+              val moshiAccess: IrExpression =
+                if (hasObjectSubtypes) {
+                  val initial =
+                    irCall(moshiSymbols.moshiNewBuilder).apply {
+                      dispatchReceiver = irGet(moshiParam)
+                    }
+                  val newBuilds =
+                    objectSubtypes.fold(initial) { receiver, subtype ->
+                      irCall(moshiSymbols.addAdapter).apply {
+                        extensionReceiver = receiver
+                        putTypeArgument(0, subtype.defaultType)
+                        putValueArgument(
+                          0,
+                          irCall(moshiSealedSymbols.objectJsonAdapterCtor).apply {
+                            putValueArgument(0, irGetObject(subtype.symbol))
+                          }
+                        )
+                      }
+                    }
+                  irCall(moshiSymbols.moshiBuilderBuild).apply { dispatchReceiver = newBuilds }
+                } else {
+                  irGet(moshiParam)
+                }
+
               val subtypesExpression =
                 subtypes.filterIsInstance<Subtype.ClassType>().fold(ofCreatorExpression) {
                   receiver,
@@ -366,45 +460,29 @@ private constructor(
                   }
                 }
 
-              val possiblyWithDefaultExpression =
-                if (useDefaultNull) {
-                  irCall(moshiSealedSymbols.pjafWithDefaultValue).apply {
-                    dispatchReceiver = subtypesExpression
-                    putValueArgument(0, irNull(targetType.defaultType))
-                  }
+              var finalFallbackStrategy = fallbackStrategy
+              subtypes.filterIsInstance<Subtype.ObjectType>().firstOrNull()?.let { defaultObject ->
+                if (fallbackStrategy == null) {
+                  finalFallbackStrategy =
+                    FallbackStrategy.DefaultObject(defaultObject.className.symbol)
                 } else {
-                  subtypes.filterIsInstance<Subtype.ObjectType>().firstOrNull()?.let { defaultObject
-                    ->
-                    irCall(moshiSealedSymbols.pjafWithDefaultValue).apply {
-                      dispatchReceiver = subtypesExpression
-                      putValueArgument(0, irGetObject(defaultObject.className.symbol))
-                    }
+                  logger.error(target) {
+                    "Only one of @DefaultObject, @DefaultNull, or @FallbackJsonAdapter can be used at a time."
                   }
-                    ?: subtypesExpression
                 }
+              }
 
-              val moshiParam = irGet(ctor.valueParameters[0])
-              val moshiAccess: IrExpression =
-                if (hasObjectSubtypes) {
-                  val initial =
-                    irCall(moshiSymbols.moshiNewBuilder).apply { dispatchReceiver = moshiParam }
-                  val newBuilds =
-                    objectSubtypes.fold(initial) { receiver, subtype ->
-                      irCall(moshiSymbols.addAdapter).apply {
-                        extensionReceiver = receiver
-                        putTypeArgument(0, subtype.defaultType)
-                        putValueArgument(
-                          0,
-                          irCall(moshiSealedSymbols.objectJsonAdapterCtor).apply {
-                            putValueArgument(0, irGetObject(subtype.symbol))
-                          }
-                        )
-                      }
-                    }
-                  irCall(moshiSymbols.moshiBuilderBuild).apply { dispatchReceiver = newBuilds }
-                } else {
-                  moshiParam
+              val possiblyWithDefaultExpression =
+                finalFallbackStrategy?.let {
+                  it.statement(
+                    builder = this,
+                    moshiSealedSymbols = moshiSealedSymbols,
+                    subtypesExpression = subtypesExpression,
+                    targetType = targetType.defaultType,
+                    moshiParam = moshiParam
+                  )
                 }
+                  ?: subtypesExpression
 
               // .create(Message::class.java, emptySet(), moshi) as JsonAdapter<Message>
               irExprBody(
@@ -564,4 +642,72 @@ private constructor(
 private sealed class Subtype(val className: IrClass) {
   class ObjectType(className: IrClass) : Subtype(className)
   class ClassType(className: IrClass, val labels: List<String>) : Subtype(className)
+}
+
+private sealed interface FallbackStrategy {
+  fun statement(
+    builder: IrBuilderWithScope,
+    moshiSealedSymbols: MoshiSealedSymbols,
+    subtypesExpression: IrExpression,
+    targetType: IrType,
+    moshiParam: IrValueParameter,
+  ): IrCall
+
+  object Null : FallbackStrategy {
+    override fun statement(
+      builder: IrBuilderWithScope,
+      moshiSealedSymbols: MoshiSealedSymbols,
+      subtypesExpression: IrExpression,
+      targetType: IrType,
+      moshiParam: IrValueParameter,
+    ) =
+      with(builder) {
+        irCall(moshiSealedSymbols.pjafWithDefaultValue).apply {
+          dispatchReceiver = subtypesExpression
+          putValueArgument(0, irNull(targetType))
+        }
+      }
+  }
+
+  class FallbackAdapter(val targetConstructor: IrFunctionSymbol, val hasMoshiParam: Boolean) :
+    FallbackStrategy {
+    override fun statement(
+      builder: IrBuilderWithScope,
+      moshiSealedSymbols: MoshiSealedSymbols,
+      subtypesExpression: IrExpression,
+      targetType: IrType,
+      moshiParam: IrValueParameter,
+    ): IrCall {
+      return with(builder) {
+        irCall(moshiSealedSymbols.pjafWithFallbackJsonAdapter).apply {
+          dispatchReceiver = subtypesExpression
+          // TODO cast to JsonAdapter<Any>
+          val args =
+            if (hasMoshiParam) {
+              arrayOf(irGet(moshiParam))
+            } else {
+              emptyArray()
+            }
+          putValueArgument(0, irInvoke(callee = targetConstructor, args = args))
+        }
+      }
+    }
+  }
+
+  class DefaultObject(val defaultObject: IrClassSymbol) : FallbackStrategy {
+    override fun statement(
+      builder: IrBuilderWithScope,
+      moshiSealedSymbols: MoshiSealedSymbols,
+      subtypesExpression: IrExpression,
+      targetType: IrType,
+      moshiParam: IrValueParameter,
+    ): IrCall {
+      return with(builder) {
+        irCall(moshiSealedSymbols.pjafWithDefaultValue).apply {
+          dispatchReceiver = subtypesExpression
+          putValueArgument(0, irGetObject(defaultObject))
+        }
+      }
+    }
+  }
 }