+++ moshi-sealed/codegen/src/main/kotlin/dev/zacsweers/moshix/sealed/codegen/ksp/MoshiSealedSymbolProcessorProvider.kt
@@ -17,6 +17,7 @@ package dev.zacsweers.moshix.sealed.codegen.ksp
 
 import com.google.auto.service.AutoService
 import com.google.devtools.ksp.getAllSuperTypes
+import com.google.devtools.ksp.isVisibleFrom
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.Dependencies
 import com.google.devtools.ksp.processing.Resolver
@@ -30,20 +31,18 @@ import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.symbol.KSFile
 import com.google.devtools.ksp.symbol.KSType
 import com.google.devtools.ksp.symbol.Modifier
+import com.squareup.kotlinpoet.ANY
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.MemberName
+import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.ksp.addOriginatingKSFile
-import com.squareup.kotlinpoet.ksp.originatingKSFiles
 import com.squareup.kotlinpoet.ksp.toClassName
 import com.squareup.kotlinpoet.ksp.writeTo
+import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.JsonClass
-import dev.zacsweers.moshix.sealed.annotations.DefaultNull
-import dev.zacsweers.moshix.sealed.annotations.DefaultObject
-import dev.zacsweers.moshix.sealed.annotations.NestedSealed
-import dev.zacsweers.moshix.sealed.annotations.TypeLabel
 import dev.zacsweers.moshix.sealed.codegen.ProguardConfig
 import dev.zacsweers.moshix.sealed.codegen.ksp.MoshiSealedSymbolProcessorProvider.Companion.OPTION_GENERATED
 import dev.zacsweers.moshix.sealed.codegen.ksp.MoshiSealedSymbolProcessorProvider.Companion.OPTION_GENERATE_PROGUARD_RULES
@@ -142,29 +141,21 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
           .build()
       }
 
-    val jsonClassType =
-      resolver.getClassDeclarationByName(resolver.getKSNameFromString(JSON_CLASS_NAME))?.asType()
-        ?: run {
-          logger.error("JsonClass type not found on the classpath.")
-          return emptyList()
-        }
-
-    val nestedSealedType = resolver.getClassDeclarationByName<NestedSealed>().asType()
-    val jsonClassAnnotation = resolver.getClassDeclarationByName<JsonClass>().asType()
+    val symbols = MoshiSealedSymbols(resolver)
 
     resolver.getSymbolsWithAnnotation(JSON_CLASS_NAME).asSequence().forEach { type ->
       check(type is KSClassDeclaration) { "@JsonClass is only applicable to classes!" }
 
-      val labelKey = type.findAnnotationWithType(jsonClassType)?.labelKey() ?: return@forEach
+      val labelKey = type.findAnnotationWithType(symbols.jsonClass)?.labelKey() ?: return@forEach
 
       check(Modifier.SEALED in type.modifiers) { "Must be a sealed class!" }
 
       // If this is a nested sealed type of a moshi-sealed parent, defer to the parent
       val sealedParent =
-        if (type.hasAnnotation(nestedSealedType)) {
+        if (type.hasAnnotation(symbols.nestedSealed)) {
           type.getAllSuperTypes().firstNotNullOfOrNull { supertype ->
             // Weird that we need to check the classifier ourselves
-            supertype.declaration.findAnnotationWithType(jsonClassAnnotation)?.labelKey()?.let {
+            supertype.declaration.findAnnotationWithType(symbols.jsonClass)?.labelKey()?.let {
               supertype to it
             }
           }
@@ -187,11 +178,8 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
         }
       }
 
-      createType(resolver, type, labelKey, generatedAnnotation, jsonClassAnnotation)
+      createType(type, labelKey, generatedAnnotation, symbols)
     }
-
-    // TODO sealedParent gen?
-    //  Requires a runtime adapter! Or JsonClass(generator = "sealed-nested")
 
     return emptyList()
   }
@@ -211,25 +199,77 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
   }
 
   private fun createType(
-    resolver: Resolver,
     type: KSClassDeclaration,
     labelKey: String,
     generatedAnnotation: AnnotationSpec?,
-    jsonClassAnnotation: KSType
+    symbols: MoshiSealedSymbols,
   ) {
-    val defaultNullAnnotation = resolver.getClassDeclarationByName<DefaultNull>().asType()
-    val defaultObjectAnnotation = resolver.getClassDeclarationByName<DefaultObject>().asType()
-    val typeLabelAnnotation = resolver.getClassDeclarationByName<TypeLabel>().asType()
-    val useDefaultNull = type.hasAnnotation(defaultNullAnnotation)
+    val useDefaultNull = type.hasAnnotation(symbols.defaultNull)
+    val fallbackAdapterAnnotation = type.findAnnotationWithType(symbols.fallbackJsonAdapter)
+    if (useDefaultNull && (fallbackAdapterAnnotation != null)) {
+      logger.error("Only one of @DefaultNull or @FallbackJsonAdapter can be used at a time", type)
+      return
+    }
+    var fallbackStrategy: FallbackStrategy? = null
+    if (fallbackAdapterAnnotation != null) {
+      val adapterType = (fallbackAdapterAnnotation.arguments[0].value as KSType)
+      // TODO can we check adapter type is valid? Compiler will check it for us
+      val adapterDeclaration = adapterType.declaration as KSClassDeclaration
+      val constructor = adapterDeclaration.primaryConstructor
+      if (constructor?.isVisibleFrom(type) == false) {
+        logger.error(
+          "Fallback adapter type $adapterType and its primary constructor must be visible from $type",
+          fallbackAdapterAnnotation
+        )
+        return
+      }
+      val hasMoshiParam =
+        when (constructor?.parameters?.size) {
+          null,
+          0 -> {
+            // Nothing to do
+            false
+          }
+          1 -> {
+            // Check it's a Moshi parameter
+            val moshiParam = constructor.parameters[0]
+            // TODO can this be simpler?
+            if (!symbols.moshi.isAssignableFrom(moshiParam.type.resolve())) {
+              logger.error(
+                "Fallback adapter type's primary constructor can only have a Moshi parameter",
+                fallbackAdapterAnnotation
+              )
+              return
+            }
+            true
+          }
+          else -> {
+            logger.error(
+              "Fallback adapter type's primary constructor can only have a Moshi parameter",
+              fallbackAdapterAnnotation
+            )
+            return
+          }
+        }
+      fallbackStrategy =
+        FallbackStrategy.FallbackAdapter(
+          className = adapterType.toClassName(),
+          hasMoshiParam = hasMoshiParam
+        )
+    } else if (useDefaultNull) {
+      fallbackStrategy = FallbackStrategy.Null
+    }
+
     val objectAdapters = mutableListOf<CodeBlock>()
     val seenLabels = mutableMapOf<String, ClassName>()
     val originatingKSFiles = mutableSetOf<KSFile>()
+    val nestedSealedClassNames = mutableSetOf<ClassName>()
     type.containingFile?.let(originatingKSFiles::add)
     val sealedSubtypes =
       type.getSealedSubclasses().flatMapTo(LinkedHashSet()) { subtype ->
         val className = subtype.toClassName()
         val isObject = subtype.classKind == OBJECT
-        if (isObject && subtype.hasAnnotation(defaultObjectAnnotation)) {
+        if (isObject && subtype.hasAnnotation(symbols.defaultObject)) {
           if (useDefaultNull) {
             // Print both for reference
             logger.error(
@@ -248,56 +288,57 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
           walkTypeLabels(
             rootType = type,
             subtype = subtype,
-            typeLabelAnnotation = typeLabelAnnotation,
-            jsonClassAnnotation = jsonClassAnnotation,
+            symbols = symbols,
             labelKey = labelKey,
             seenLabels = seenLabels,
             objectAdapters = objectAdapters,
             originatingKSFiles = originatingKSFiles,
+            nestedSealedClassNames = nestedSealedClassNames,
             className = className
           )
         }
       }
 
-    val preparedAdapter =
-      createType(
+    createType(
         targetType = type.toClassName(),
         isInternal = Modifier.INTERNAL in type.modifiers,
         labelKey = labelKey,
-        useDefaultNull = useDefaultNull,
+        fallbackStrategy = fallbackStrategy,
         generatedAnnotation = generatedAnnotation,
+        nestedSealedClassNames = nestedSealedClassNames,
         subtypes = sealedSubtypes,
         objectAdapters = objectAdapters,
-        generateProguardConfig = generateProguardConfig
+        generateProguardConfig = generateProguardConfig,
+        errorLogger = { message -> logger.error(message, type) }
       ) {
         addAnnotation(COMMON_SUPPRESS)
         for (file in originatingKSFiles) {
           addOriginatingKSFile(file)
         }
       }
-
-    preparedAdapter.spec.writeTo(codeGenerator, aggregating = true)
-    preparedAdapter.proguardConfig?.writeTo(codeGenerator, type.containingFile)
+      ?.let { (spec, proguardConfig) ->
+        spec.writeTo(codeGenerator, aggregating = true)
+        proguardConfig?.writeTo(codeGenerator, type.containingFile)
+      }
   }
 
   private fun walkTypeLabels(
     rootType: KSClassDeclaration,
     subtype: KSClassDeclaration,
-    typeLabelAnnotation: KSType,
-    jsonClassAnnotation: KSType,
+    symbols: MoshiSealedSymbols,
     labelKey: String,
     seenLabels: MutableMap<String, ClassName>,
     objectAdapters: MutableList<CodeBlock>,
     originatingKSFiles: MutableSet<KSFile>,
+    nestedSealedClassNames: MutableSet<ClassName>,
     className: ClassName = subtype.toClassName(),
   ): Sequence<Subtype> {
     subtype.containingFile?.let(originatingKSFiles::add)
     // If it's sealed, check if it's inheriting from our existing type or a separate/new branching
-    // off
-    // point
+    // off point
     if (Modifier.SEALED in subtype.modifiers) {
       val nestedLabelKey =
-        subtype.findAnnotationWithType(jsonClassAnnotation)?.labelKey(checkGenerateAdapter = false)
+        subtype.findAnnotationWithType(symbols.jsonClass)?.labelKey(checkGenerateAdapter = false)
       if (nestedLabelKey != null) {
         // Redundant case
         if (labelKey == nestedLabelKey) {
@@ -308,14 +349,13 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
         }
       }
 
-      if (subtype.findAnnotationWithType(typeLabelAnnotation) != null) {
+      if (subtype.findAnnotationWithType(symbols.typeLabel) != null) {
         // It's a different type, allow it to be used as a label and branch off from here.
         val classType =
           addLabelKeyForType(
             rootType,
             subtype,
-            typeLabelAnnotation,
-            jsonClassAnnotation,
+            symbols,
             seenLabels,
             objectAdapters,
             className,
@@ -323,17 +363,24 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
           )
         return classType?.let { sequenceOf(it) } ?: emptySequence()
       } else {
+        // Note this is an intermediate nested sealed type
+        if (subtype.hasAnnotation(symbols.nestedSealed)) {
+          nestedSealedClassNames += className
+        }
+        // Add the file as an originating element as it's indirectly participating in adapter
+        // generation
+        originatingKSFiles += subtype.containingFile!!
         // Recurse, inheriting the top type
         return subtype.getSealedSubclasses().flatMap {
           walkTypeLabels(
             rootType = rootType,
             subtype = it,
-            typeLabelAnnotation = typeLabelAnnotation,
-            jsonClassAnnotation = jsonClassAnnotation,
+            symbols = symbols,
             labelKey = labelKey,
             seenLabels = seenLabels,
             objectAdapters = objectAdapters,
-            originatingKSFiles = originatingKSFiles
+            originatingKSFiles = originatingKSFiles,
+            nestedSealedClassNames = nestedSealedClassNames,
           )
         }
       }
@@ -342,8 +389,7 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
         addLabelKeyForType(
           rootType = rootType,
           subtype = subtype,
-          typeLabelAnnotation = typeLabelAnnotation,
-          jsonClassAnnotation = jsonClassAnnotation,
+          symbols = symbols,
           seenLabels = seenLabels,
           objectAdapters = objectAdapters,
           className = className
@@ -355,8 +401,7 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
   private fun addLabelKeyForType(
     rootType: KSClassDeclaration,
     subtype: KSClassDeclaration,
-    typeLabelAnnotation: KSType,
-    jsonClassAnnotation: KSType,
+    symbols: MoshiSealedSymbols,
     seenLabels: MutableMap<String, ClassName>,
     objectAdapters: MutableList<CodeBlock>,
     className: ClassName = subtype.toClassName(),
@@ -364,7 +409,7 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
   ): Subtype? {
     // Regular subtype, read its label
     val labelAnnotation =
-      subtype.findAnnotationWithType(typeLabelAnnotation)
+      subtype.findAnnotationWithType(symbols.typeLabel)
         ?: run {
           logger.error("Missing @TypeLabel", subtype)
           return null
@@ -391,12 +436,12 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
 
     labels += mainLabel
 
+    // https://github.com/google/ksp/issues/134
     @Suppress("UNCHECKED_CAST")
     val alternates =
       labelAnnotation.arguments.find { it.name?.getShortName() == "alternateLabels" }?.value
         as? List<String> // arrays are lists in KSP https://github.com/google/ksp/issues/135
        ?: emptyList() // ksp ignores undefined args
-    // https://github.com/google/ksp/issues/134
 
     for (alternate in alternates) {
       seenLabels.put(alternate, className)?.let { prev ->
@@ -409,7 +454,7 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
     }
 
     if (!skipJsonClassCheck) {
-      val labelKey = subtype.findAnnotationWithType(jsonClassAnnotation)?.labelKey()
+      val labelKey = subtype.findAnnotationWithType(symbols.jsonClass)?.labelKey()
       if (labelKey != null) {
         logger.error(
           "Sealed subtype $subtype is annotated with @JsonClass(generator = \"sealed:$labelKey\") and @TypeLabel.",
@@ -433,6 +478,37 @@ private class MoshiSealedSymbolProcessor(environment: SymbolProcessorEnvironment
     }
 
     return Subtype.ClassType(className, labels)
+  }
+}
+
+internal sealed interface FallbackStrategy {
+  fun statement(moshiParam: CodeBlock): CodeBlock
+
+  object Null : FallbackStrategy {
+    override fun statement(moshiParam: CodeBlock) = CodeBlock.of(".withDefaultValue(null)")
+  }
+
+  class FallbackAdapter(val className: TypeName, val hasMoshiParam: Boolean) : FallbackStrategy {
+    // TODO handle which moshi param comes in here
+    override fun statement(moshiParam: CodeBlock): CodeBlock {
+      val constructorParams =
+        if (hasMoshiParam) {
+          moshiParam
+        } else {
+          CodeBlock.of("")
+        }
+      return CodeBlock.of(
+        ".withFallbackJsonAdapter(%T(%L)·as·%T<%T>)",
+        className,
+        constructorParams,
+        JsonAdapter::class.asClassName(),
+        ANY
+      )
+    }
+  }
+
+  class DefaultObject(val className: TypeName) : FallbackStrategy {
+    override fun statement(moshiParam: CodeBlock) = CodeBlock.of(".withDefaultValue(%T)", className)
   }
 }