+++ moshi-sealed/metadata-reflect/src/main/kotlin/dev/zacsweers/moshix/sealed/reflect/MetadataMoshiSealedJsonAdapterFactory.kt
@@ -22,15 +22,17 @@ import com.squareup.moshi.adapters.PolymorphicJsonAdapterFactory
 import com.squareup.moshi.rawType
 import dev.zacsweers.moshix.sealed.annotations.DefaultNull
 import dev.zacsweers.moshix.sealed.annotations.DefaultObject
+import dev.zacsweers.moshix.sealed.annotations.FallbackJsonAdapter
 import dev.zacsweers.moshix.sealed.annotations.NestedSealed
 import dev.zacsweers.moshix.sealed.annotations.TypeLabel
 import dev.zacsweers.moshix.sealed.runtime.internal.ObjectJsonAdapter
+import dev.zacsweers.moshix.sealed.runtime.internal.Util.fallbackAdapter
 import java.lang.reflect.Type
 import kotlinx.metadata.ClassName
 import kotlinx.metadata.Flag
 import kotlinx.metadata.KmClass
-import kotlinx.metadata.jvm.KotlinClassHeader
 import kotlinx.metadata.jvm.KotlinClassMetadata
+import kotlinx.metadata.jvm.Metadata
 
 /** Classes annotated with this are eligible for this adapter. */
 private val KOTLIN_METADATA = Metadata::class.java
@@ -75,6 +77,13 @@ public class MetadataMoshiSealedJsonAdapterFactory : JsonAdapter.Factory {
       if (rawType.isAnnotationPresent(DefaultNull::class.java)) {
         defaultObjectInstance = null
       }
+      var fallbackAdapter: JsonAdapter<Any>? = null
+      val fallbackJsonAdapterAnnotation = rawType.getAnnotation(FallbackJsonAdapter::class.java)
+      if (fallbackJsonAdapterAnnotation != null) {
+        val clazz = fallbackJsonAdapterAnnotation.value
+        // Find a constructor we can use
+        fallbackAdapter = moshi.fallbackAdapter(clazz.java)
+      }
 
       val objectSubtypes = mutableMapOf<Class<*>, Any>()
       val labels = mutableMapOf<String, Class<*>>()
@@ -87,11 +96,13 @@ public class MetadataMoshiSealedJsonAdapterFactory : JsonAdapter.Factory {
         if (isAnnotatedDefaultObject) {
           if (!isObject) {
             error("Must be an object type to use as a @DefaultObject: $sealedSubclass")
-          } else if (defaultObjectInstance === UNSET) {
+          } else if (defaultObjectInstance === UNSET && fallbackAdapter == null) {
             defaultObjectInstance = sealedSubclass.objectInstance()
           } else {
-            if (defaultObjectInstance == null) {
-              error("Can not have both @DefaultObject and @DefaultNull: $sealedSubclass")
+            if (defaultObjectInstance == null || fallbackAdapter == null) {
+              error(
+                "Only one of @DefaultNull, @DefaultObject, and @FallbackJsonAdapter: $sealedSubclass"
+              )
             } else {
               error(
                 "Can only have one @DefaultObject: $sealedSubclass and ${defaultObjectInstance.javaClass} are both annotated"
@@ -128,6 +139,8 @@ public class MetadataMoshiSealedJsonAdapterFactory : JsonAdapter.Factory {
           .let { factory ->
             if (defaultObjectInstance !== UNSET) {
               factory.withDefaultValue(defaultObjectInstance)
+            } else if (fallbackAdapter != null) {
+              factory.withFallbackJsonAdapter(fallbackAdapter)
             } else {
               factory
             }
@@ -140,10 +153,10 @@ public class MetadataMoshiSealedJsonAdapterFactory : JsonAdapter.Factory {
   }
 }
 
-private fun Class<*>.header(): KotlinClassHeader? {
+private fun Class<*>.header(): Metadata? {
   val metadata = getAnnotation(KOTLIN_METADATA) ?: return null
   return with(metadata) {
-    KotlinClassHeader(
+    Metadata(
       kind = kind,
       metadataVersion = metadataVersion,
       data1 = data1,
@@ -155,7 +168,7 @@ private fun Class<*>.header(): KotlinClassHeader? {
   }
 }
 
-private fun KotlinClassHeader.toKmClass(): KmClass? {
+private fun Metadata.toKmClass(): KmClass? {
   val classMetadata = KotlinClassMetadata.read(this)
   if (classMetadata !is KotlinClassMetadata.Class) {
     return null